Ensure actual message size at least equals StructureSize.
It could sometimes be one byte smaller, because the variable "buffer" part of the message could be empty (e.g. for a read, or for operations on the root directory of the share). This caused Windows servers to give an error, although macOS was OK with it. We now add a byte of padding to avoid the error from Windows.