Add Finder extension to prevent it from crashing on long names.
Previously, the Finder would display garbage at the end of names longer than 32 characters, and it would crash if it encountered a very long name. This extension patches it to prevent those things, although the Finder still doesn't keep track of characters beyond the 32nd internally, and therefore won't be able to open or do some other operations on files with longer names.