Implement reconnection for open files.
This does not use the SMB2 durable handle features; it just re-opens the files.

To guard against cases where the file may have been deleted or moved and then another file created with the same name, we require the creation date to match what it was originally. This is not a totally reliable check (since creation dates can be changed), but it should work in most cases.