Cache EOF locally to avoid going to server on every SetMark call.
This essentially cuts the amount of network traffic in half when doing reads or writes with a random-access pattern (e.g. accessing resources).

The cached EOF value should remain correct if the server fully enforces the sharing rules, but if not it may become incorrect due to file changes on the server. Therefore, we only use the cached EOF to check whether to give an outOfRange error from SetMark, and even then we check the EOF on the server before reporting the error. If the file has shrunk on the server, it is possible that we may not detect a condition where we should report outOfRange, but that is unlikely to cause problems in practice (an error may occur on a subsequent read or write instead, but that might have happened anyway if the timing was different).