Reconnection enhancements.
The SMB FST can now reconnect if the server reports that the session is deleted or expired (which Windows sometimes does). It can also reconnect when logging in (if the session is dropped because the user takes a while to enter their username/password). There are also some additional safeguards to prevent reconnection at inappropriate times.