Fix logic for checking ReplyBitmap in AAPL response.
The old logic was buggy and effectively ignored ReplyBitmap. But just fixing it to work as originally intended would have caused the AAPL support in illumos to be ignored, because illumos does not include the model string in its AAPL response. The logic has now been fixed to address this issue, reflecting the fact that the server capabilities are all we really need (although we will look at the model string if present to help differentiate Samba from actual Macs).