Allow file creation on older Samba servers without streams support.
Older Samba versions give a different error code than Windows and current Samba. Since it does not unambiguously identify that lack of streams support is the issue, we also verify that by querying the filesystem attributes.

As on other platforms without streams support, file types will not be set correctly, but at least files can be created and written to.