Reduce read/write buffer sizes when available memory is low.
Marinetti needs some free memory to operate properly, and it can sometimes crash if memory allocation fails. To try to mitigate these issues, we now scale down the buffer sizes used for reads and writes if there is limited free memory. This should reduce the memory demands within Marinetti. This is not completely guaranteed to fix all issues, but in my testing it did allow read and write operations to work in low-memory situations where they would previously fail or sometimes crash.

(This can come up when doing copies in the Finder, since it reads as much data as possible into RAM before writing it out, meaning it can do reads and writes in situations where there is little free memory.)